+++ clarity/src/vm/contexts.rs
@@ -14,9 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+use std::cell::RefCell;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::fmt;
 use std::mem::replace;
+use std::rc::Rc;
 use std::time::{Duration, Instant};
 
 use clarity_types::representations::ClarityName;
@@ -197,6 +199,12 @@ impl AssetMap {
     }
 }
 
+/// EventBatchHook defines an interface for hooks to execute before and/or after each event push.
+pub trait EventBatchHook {
+    fn on_push(&mut self, event: &StacksTransactionEvent);
+    fn on_rollback(&mut self, event_batch: &EventBatch, reason: RollbackReason);
+}
+
 #[derive(Debug, Clone, Default)]
 pub struct EventBatch {
     pub events: Vec<StacksTransactionEvent>,
@@ -232,6 +240,8 @@ pub struct GlobalContext<'a, 'hooks> {
     pub chain_id: u32,
     pub eval_hooks: Option<Vec<&'hooks mut dyn EvalHook>>,
     pub execution_time_tracker: ExecutionTimeTracker,
+    /// hooks to run at every event push
+    pub event_batch_hooks: Option<Vec<Rc<RefCell<dyn EventBatchHook>>>>,
 }
 
 #[derive(Serialize, Deserialize, Clone)]
@@ -708,7 +718,7 @@ impl<'a, 'hooks> OwnedEnvironment<'a, 'hooks> {
                 Ok((return_value, asset_map, event_batch.events))
             }
             Err(e) => {
-                self.context.roll_back()?;
+                self.context.roll_back(RollbackReason::VmError)?;
                 Err(e)
             }
         }
@@ -1047,7 +1057,7 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
             .database
             .get_contract(contract_identifier)
             .or_else(|e| {
-                self.global_context.roll_back()?;
+                self.global_context.roll_back(RollbackReason::VmError)?;
                 Err(e)
             })?;
 
@@ -1064,7 +1074,11 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
             eval(&parsed[0], &mut nested_env, &local_context)
         };
 
-        self.global_context.roll_back()?;
+        self.global_context.roll_back(if result.is_ok() {
+            RollbackReason::Readonly
+        } else {
+            RollbackReason::VmError
+        })?;
 
         result
     }
@@ -1262,7 +1276,18 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
         };
 
         if make_read_only {
-            self.global_context.roll_back()?;
+            let reason = match &result {
+                Ok(Value::Response(response_data)) => {
+                    if response_data.committed {
+                        RollbackReason::Readonly
+                    } else {
+                        RollbackReason::VmError
+                    }
+                }
+                Ok(_) => RollbackReason::Readonly,
+                Err(_) => RollbackReason::VmError,
+            };
+            self.global_context.roll_back(reason)?;
             result
         } else {
             self.global_context.handle_tx_result(result, allow_private)
@@ -1294,7 +1319,11 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
                 result
             });
 
-        self.global_context.roll_back()?;
+        self.global_context.roll_back(if result.is_ok() {
+            RollbackReason::Readonly
+        } else {
+            RollbackReason::VmError
+        })?;
 
         result
     }
@@ -1383,7 +1412,7 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
                 Ok(())
             }
             Err(e) => {
-                self.global_context.roll_back()?;
+                self.global_context.roll_back(RollbackReason::VmError)?;
                 Err(e)
             }
         }
@@ -1413,12 +1442,12 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
                     Ok(value)
                 }
                 Err(_) => {
-                    self.global_context.roll_back()?;
+                    self.global_context.roll_back(RollbackReason::VmError)?;
                     Err(VmInternalError::InsufficientBalance.into())
                 }
             },
             Err(e) => {
-                self.global_context.roll_back()?;
+                self.global_context.roll_back(RollbackReason::VmError)?;
                 Err(e)
             }
         }
@@ -1437,7 +1466,7 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
                 Ok(ret)
             }
             Err(e) => {
-                self.global_context.roll_back()?;
+                self.global_context.roll_back(RollbackReason::VmError)?;
                 Err(e)
             }
         }
@@ -1455,7 +1484,14 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
             0
         };
         if let Some((batch, total_size)) = self.global_context.event_batches.last_mut() {
+            if let Some(event_batch_hooks) = &mut self.global_context.event_batch_hooks {
+                event_batch_hooks
+                    .iter_mut()
+                    .for_each(|event_batch_hook| event_batch_hook.borrow_mut().on_push(&event));
+            }
+
             batch.events.push(event);
+
             *total_size = total_size.saturating_add(size.into());
             if *total_size >= MAX_EVENTS_BATCH {
                 return Err(VmInternalError::Expect(
@@ -1627,6 +1663,14 @@ impl<'a, 'b, 'hooks> Environment<'a, 'b, 'hooks> {
     }
 }
 
+#[derive(Clone, Debug)]
+pub enum RollbackReason {
+    // normal rollback for readonly calls
+    Readonly,
+    // caused by an error
+    VmError,
+}
+
 impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
     // Instantiate a new Global Context
     pub fn new(
@@ -1647,6 +1691,7 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
             chain_id,
             eval_hooks: None,
             execution_time_tracker: ExecutionTimeTracker::NoTracking,
+            event_batch_hooks: None,
         }
     }
 
@@ -1735,7 +1780,7 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
     {
         self.begin();
         let result = f(self).or_else(|e| {
-            self.roll_back()?;
+            self.roll_back(RollbackReason::VmError)?;
             Err(e)
         })?;
         self.commit()?;
@@ -1772,7 +1817,11 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
             );
             f(&mut exec_env)
         };
-        self.roll_back()?;
+        self.roll_back(if result.is_ok() {
+            RollbackReason::Readonly
+        } else {
+            RollbackReason::VmError
+        })?;
 
         match result {
             Ok(return_value) => Ok(return_value),
@@ -1844,7 +1893,7 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
         Ok((out_map, out_batch))
     }
 
-    pub fn roll_back(&mut self) -> Result<(), VmExecutionError> {
+    pub fn roll_back(&mut self, reason: RollbackReason) -> Result<(), VmExecutionError> {
         let popped = self.asset_maps.pop();
         if popped.is_none() {
             return Err(VmInternalError::Expect("Expected entry to rollback".into()).into());
@@ -1853,8 +1902,17 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
         if popped.is_none() {
             return Err(VmInternalError::Expect("Expected entry to rollback".into()).into());
         }
+
         let popped = self.event_batches.pop();
-        if popped.is_none() {
+        if let Some((event_batch, _size)) = popped {
+            if let Some(event_batch_hooks) = &mut self.event_batch_hooks {
+                event_batch_hooks.iter_mut().for_each(|event_batch_hook| {
+                    event_batch_hook
+                        .borrow_mut()
+                        .on_rollback(&event_batch, reason.clone())
+                });
+            }
+        } else {
             return Err(VmInternalError::Expect("Expected entry to rollback".into()).into());
         }
 
@@ -1874,7 +1932,7 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
                 if data.committed {
                     self.commit()?;
                 } else {
-                    self.roll_back()?;
+                    self.roll_back(RollbackReason::VmError)?;
                 }
                 Ok(Value::Response(data))
             } else if allow_private && cfg!(feature = "devtools") {
@@ -1887,7 +1945,7 @@ impl<'a, 'hooks> GlobalContext<'a, 'hooks> {
                 .into())
             }
         } else {
-            self.roll_back()?;
+            self.roll_back(RollbackReason::VmError)?;
             result
         }
     }